requirements.txt
djangorestframework>=3.14.0, <3.15.0
psycopg2-binary
requests>=2.31.0, <2.32.0
numpy
django-redis
redis
matplotlib
